AttackCommand

+++ Assets/Scripts/Unit/Stats.cs
@@ -8,6 +8,6 @@ namespace DarkLegion.Unit
 
         [field: SerializeField, Range(0f, 1000f)] public int Health { get; private set; }
 
-        [field: SerializeField, Range(0f, 100f)] public float PhysicalDamage { get; private set; }
+        [field: SerializeField, Range(0f, 100f)] public int PhysicalDamage { get; private set; }
     }
 }
